Get rid of support for separate guides for first step.
Not currently used, and getting in the way. And besides, I don't think
this is the right way to factor this. If we do want to try this, I
think we need to push this into the main part of the guide in such a
way that all of the various w and z guides can support it. e.g
image-so-far, etc.